backend/pyproject.toml
[project]
name = "autovend-backend"
version = "1.0.0"
description = "LlamaIndex-based intelligent automotive sales assistant with RAG architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LlamaIndex core
    "llama-index-core>=0.11.0",
    "llama-index-llms-openai-like>=0.2.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-vector-stores-chroma>=0.2.0",
    # Vector store
    "chromadb>=0.5.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    # Data models & validation
    "pydantic>=2.10.0",
    # Configuration
    "python-dotenv>=1.0.0",
    # Data parsing
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"agent/extractors/*.py" = ["E501"]
"agent/response_generator.py" = ["E501"]
"tests/kpi_report.py" = ["E501"]

[tool.coverage.run]
source = ["app", "agent"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 85
show_missing = true
